+++ src/app/shared/RelayActions.svelte
@@ -3,7 +3,7 @@
   import OverflowMenu from "src/partials/OverflowMenu.svelte"
   import user from "src/agent/user"
   import {getRelayWithFallback} from "src/agent/db"
-  import {addToFeed} from "src/app/state"
+  import {addToList} from "src/app/state"
 
   export let relay
 
@@ -32,8 +32,8 @@
     }
 
     actions.push({
-      onClick: () => addToFeed("relays", relay.url),
-      label: "Add to feed",
+      onClick: () => addToList("r", relay.url),
+      label: "Add to list",
       icon: "scroll",
     })