+++ src/app/views/Feeds.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import type {CustomFeed} from "src/util/types"
-  import {prop, objOf, find, propEq} from "ramda"
+  import {prop, uniq, indexBy, objOf, filter as _filter} from "ramda"
   import {shuffle, synced} from "src/util/misc"
+  import {Tags} from "src/util/nostr"
   import {modal} from "src/partials/state"
   import Anchor from "src/partials/Anchor.svelte"
   import Content from "src/partials/Content.svelte"
@@ -12,74 +12,57 @@
   import {sampleRelays, getAllPubkeyWriteRelays, getUserReadRelays} from "src/agent/relays"
   import user from "src/agent/user"
 
-  const {feeds} = user
-  const defaultFeeds = [
-    {id: "follows", name: "Follows", authors: "follows"},
-    {id: "network", name: "Network", authors: "network"},
-  ] as Array<CustomFeed>
+  const {lists} = user
+  const activeTab = synced("views/Feeds/activeTab", "Follows")
 
-  let activeTab = synced("views/Feeds/activeTab", "Follows")
-  let relays, filter, tabs, feed
+  let relays, filter, tabs
 
-  $: allFeeds = defaultFeeds.concat($feeds)
-
+  $: listsByName = indexBy(l => Tags.from(l).getMeta("d"), $lists)
   $: {
-    tabs = allFeeds.map(prop("name")).slice(0, 2)
+    const defaultTabs = ["Follows", "Network"]
+    const customTabs = Object.keys(listsByName)
+    const validTabs = defaultTabs.concat(customTabs)
 
-    if (!tabs.includes($activeTab)) {
-      tabs = tabs.concat($activeTab)
-    } else if ($feeds.length > 0) {
-      tabs = tabs.concat($feeds[0].name)
+    if (!validTabs.includes($activeTab)) {
+      $activeTab = validTabs[0]
     }
+
+    tabs = uniq(defaultTabs.concat($activeTab).concat(customTabs)).slice(0, 3)
   }
 
   $: {
-    feed = find(propEq("name", $activeTab), allFeeds)
+    if ($activeTab === "Follows") {
+      const authors = shuffle(getUserFollows()).slice(0, 256)
 
-    if (!feed) {
-      feed = allFeeds[0]
-      $activeTab = feed.name
-    }
-  }
-
-  $: {
-    let {authors, topics} = feed
-
-    if (authors === "follows") {
-      authors = shuffle(getUserFollows()).slice(0, 256)
+      filter = {authors}
       relays = sampleRelays(getAllPubkeyWriteRelays(authors))
-    } else if (authors === "network") {
-      authors = shuffle(getUserNetwork()).slice(0, 256)
+    } else if ($activeTab === "Network") {
+      const authors = shuffle(getUserNetwork()).slice(0, 256)
+
+      filter = {authors}
       relays = sampleRelays(getAllPubkeyWriteRelays(authors))
-    } else if (feed.relays) {
-      relays = feed.relays.map(objOf("url"))
     } else {
-      relays = sampleRelays(getUserReadRelays())
+      const list = listsByName[$activeTab]
+      const tags = Tags.from(list)
+      const authors = tags.type("p").values().all()
+      const topics = tags.type("t").values().all()
+      const urls = tags.type("r").values().all()
+
+      filter = _filter(prop("length"), {authors, "#t": topics})
+      relays = urls.length > 0 ? urls.map(objOf("url")) : sampleRelays(getUserReadRelays())
     }
 
     // Separate notes and reactions into two queries since otherwise reactions dominate,
     // we never find their parents (or reactions are mostly to a few posts), and the feed sucks
-    filter = [1, 7].map(kind => {
-      const filter = {kinds: [kind]} as Record<string, any>
-
-      if (authors) {
-        filter.authors = authors
-      }
-
-      if (topics) {
-        filter["#t"] = topics
-      }
-
-      return filter
-    })
+    filter = [1, 7].map(kind => ({...filter, kinds: [kind]}))
   }
 
   const setActiveTab = tab => {
     $activeTab = tab
   }
 
-  const showFeedsList = () => {
-    modal.push({type: "feed/list"})
+  const showLists = () => {
+    modal.push({type: "list/list"})
   }
 
   document.title = $activeTab
@@ -96,29 +79,32 @@
   {/if}
   <div>
     <Tabs {tabs} activeTab={$activeTab} {setActiveTab}>
-      {#if $feeds.length > 0}
+      {#if $lists.length > 1}
         <Popover placement="bottom" opts={{hideOnClick: true}} theme="transparent">
           <i slot="trigger" class="fa fa-ellipsis-v cursor-pointer p-2" />
           <div
             slot="tooltip"
             class="flex flex-col items-start rounded border border-solid border-gray-8 bg-black">
-            {#each $feeds as feed (feed.name)}
-              <button
-                class="w-full py-2 px-3 text-left hover:bg-gray-7"
-                on:click={() => {
-                  $activeTab = feed.name
-                }}>
-                <i class="fa fa-scroll fa-sm mr-1" />
-                {feed.name}
-              </button>
+            {#each $lists as e (e.id)}
+              {@const meta = Tags.from(e).asMeta()}
+              {#if meta.d !== $activeTab}
+                <button
+                  class="w-full py-2 px-3 text-left hover:bg-gray-7"
+                  on:click={() => {
+                    $activeTab = meta.d
+                  }}>
+                  <i class="fa fa-scroll fa-sm mr-1" />
+                  {meta.d}
+                </button>
+              {/if}
             {/each}
-            <button on:click={showFeedsList} class="w-full py-2 px-3 text-left hover:bg-gray-7">
+            <button on:click={showLists} class="w-full py-2 px-3 text-left hover:bg-gray-7">
               <i class="fa fa-cog fa-sm mr-1" /> Customize
             </button>
           </div>
         </Popover>
       {:else}
-        <i class="fa fa-ellipsis-v cursor-pointer p-1" on:click={showFeedsList} />
+        <i class="fa fa-ellipsis-v cursor-pointer p-1" on:click={showLists} />
       {/if}
     </Tabs>
     {#key $activeTab}